Move ParseTime func
